Flush the serial buffer
With enough hammering, the buffer gets filled with more data than it can
handle and causes problems with the MIDI output. Ideally, there 
shouldn't be a buffer here in the first place, because we're dealing 
with constant time data anyway.

Flush the buffer after sending the USB host data to the serial port.